Reorder codebase to support Swift FFI bridge. Adds a new processor module that will be used to process the input data. This will be used to create a Swift FFI bridge.
